+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_hi.dart
@@ -130,6 +130,15 @@ class StreamChatLocalizationsHi extends GlobalStreamChatLocalizations {
   String get instantCommandsLabel => 'तत्काल आदेश';
 
   @override
+  String get commandUnavailableWhileEditingError => 'Not available while editing';
+
+  @override
+  String get commandUnavailableWhileQuotingError => 'Not available while replying';
+
+  @override
+  String get commandUnavailableError => 'Command not available';
+
+  @override
   String fileTooLargeAfterCompressionError(double limitInMB) =>
       'फ़ाइल अपलोड करने के लिए बहुत बड़ी है। '
       'फ़ाइल आकार सीमा $limitInMB MB है। '